tbman: Add NumberEditView
This is stripped down version of Cursive's EditView but only allows
numeric input. The base can be configured when the NumberEditView is
created. Currently only support binary, decimal and hexadecimal formats.
All the editing is done in-place instead of having a separate submit
functionality.

This should be enough for tbman for now so take it into use in the
registers view. This way we can make both register "values" editable and
the user can use whichever is better or both.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>