tbmargin: Initialize tests in single step in run_margining()
This eliminates the need to make the vector mutable and is arguably more
readable. With vectors this also avoids the potential overhead of
reallocation and copying from multiple push() calls.

In general, the design of the language is such that writing and
refactoring code is easier when objects are kept immutable. For example
due to the way borrowing rules work. The immutable by default and single
step initialization principles aim to simplify reasoning about code by
providing a guarantee that every object visible in the scope has been
initialized, since it changing the state of an object after it's
creation is not possible.

Signed-off-by: Aapo Vienamo <aapo.vienamo@linux.intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>