margining: Replace Caps tuple with array
Use an array instead of a tuple to represent the raw capabilities data.
This is done because the upcoming register parsing changes require the
data words to be indexable with the [] operator. The array is arguably
more appropriate collection in this case since it guarantees that all
the elements have uniform type, whereas each element of a tuple can have
any type.

Signed-off-by: Aapo Vienamo <aapo.vienamo@linux.intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>