util: Add RegBit and RegField
Implement abstractions for expressing single bits and multi bit fields
in an array of register words. This aims to replace manual shifts and
masks. The bit and field descriptions are implemented as types with
const generics and thus do not incur any runtime cost in themselves and
as long as the accessor functions are inlined, they have no additional
runtime cost compared to manual shifts and masks.

Signed-off-by: Aapo Vienamo <aapo.vienamo@linux.intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>