margining: Handle independent margins with types
Encode the different types of independent results in enum types that are
returned from margining as a data structure. Process these enums
accordingly on the tbmargin side. This moves the logic and preconditions
of independent lane handling out of tbmargin and into generation of the
data structures, which makes the division of responsibilities between
margining core and tbmargin clearer, as tbmargin is only concerned with
the formatting of the received data structure.

Signed-off-by: Aapo Vienamo <aapo.vienamo@linux.intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>