scripts/tb-bandwidth.sh: Use PCI address to access i915_display_info
The device identifier (e.g., ~0~ or ~1~) for accessing i915_display_info
is not consistent across systems, as it depends on the order of GPU
initialization.

To make the script more robust and general, switch to using the PCI
address, which uniquely identifies the GPU device and remains stable
regardless of initialization order.

Cc: Imre Deak <imre.deak@intel.com>
Cc: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Khaled Almahallawy <khaled.almahallawy@intel.com>